Получение преподавателя по id
